# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/revenue/plans": {
      "get": {
        "operationId": "RevenueController_listPlans",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "post": {
        "operationId": "RevenueController_createPlan",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/plans/{id}": {
      "get": {
        "operationId": "RevenueController_findPlan",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "put": {
        "operationId": "RevenueController_replacePlan",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "patch": {
        "operationId": "RevenueController_updatePlan",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "delete": {
        "operationId": "RevenueController_removePlan",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/addons": {
      "get": {
        "operationId": "RevenueController_listAddons",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "post": {
        "operationId": "RevenueController_createAddon",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/addons/{id}": {
      "get": {
        "operationId": "RevenueController_findAddon",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "put": {
        "operationId": "RevenueController_replaceAddon",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "patch": {
        "operationId": "RevenueController_updateAddon",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "delete": {
        "operationId": "RevenueController_removeAddon",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/subscriptions": {
      "get": {
        "operationId": "RevenueController_listSubscriptions",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "post": {
        "operationId": "RevenueController_createSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/subscriptions/{id}": {
      "get": {
        "operationId": "RevenueController_findSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "put": {
        "operationId": "RevenueController_replaceSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "patch": {
        "operationId": "RevenueController_updateSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      },
      "delete": {
        "operationId": "RevenueController_removeSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/subscriptions/upgrade": {
      "post": {
        "operationId": "RevenueController_upgradeSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/subscriptions/{id}/cancel": {
      "post": {
        "operationId": "RevenueController_cancelSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
    "/api/revenue/subscriptions/{id}/renew": {
      "post": {
        "operationId": "RevenueController_renewSubscription",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
...
    "/api/revenue/overview": {
      "get": {
        "operationId": "RevenueController_getOverview",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full management; COO for organization billing.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "revenue"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rorizon Review-4 API",
    "description": "NestJS in-memory backend with RBAC role headers for Rorizon workflows.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role for RBAC authorization. Allowed values: Super User, COO, Manager, Analyst."
      }
    },
    "schemas": {
...
      "CreateAddonDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Advanced Analytics Module"
          },
          "description": {
            "type": "string",
            "example": "Deep-dive Scope 1-3 analytics and AI forecasting."
          },
          "priceMonthly": {
            "type": "number",
            "example": 1500
          },
          "priceAnnual": {
            "type": "number",
            "example": 15000
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "name",
          "priceMonthly",
          "priceAnnual"
        ]
      },
...
    }
  }
}